Added Complete Google OAuth Support

+++ controllers/api/v1/posts_api.js
@@ -23,13 +23,19 @@ module.exports.index = async function (req, res) {
 module.exports.destroy = async function (req, res) {
   try {
     let post = await Post.findById(req.params.id);
+    if (post.user  == req.user.id) {
       post.remove();
 
       await Comment.deleteMany({ post: req.params.id });
 
-      return res.json(200 , {
-        message : "Post and Associated Comments Deleted"
+      return res.json(200, {
+        message: "Post and Associated Comments Deleted",
       });
+    } else {
+      return res.json(401, {
+        message: "You cannot delete the Post",
+      });
+    }
   } catch (err) {
     console.log("*****" , err)
     return res.json(500 , {
